% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fugazi.R
\docType{data}
\name{fugazi}
\alias{fugazi}
\title{Toy dataset of Music Taste and Age.}
\format{
A data frame with 90 rows and 3 variables:
\describe{
\item{music}{Music Type: Fugazi, Abba, and Manilow}
\item{age}{Age Group: 0-40 or 40+}
\item{liking}{Liking Rating: Scale with a range from -118 to 122}
...
}
}
\source{
internally created toy dataset
}
\usage{
fugazi
}
\description{
A dataset containing the music type, age group, and liking rating.
}
\keyword{datasets}
