% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happyparent.R
\docType{data}
\name{happyparent}
\alias{happyparent}
\title{Toy dataset of Parents Happiness.}
\format{
A data frame with 302 rows and 7 variables:
\describe{
\item{sex}{Sex of the parent: male and female}
\item{health}{Self Rated Physical Health: 1-5}
\item{educ}{Educational Background: Ranging from 8th Grade to PhD}
\item{str_slf1}{Stress Related to Self: Ranging from 0 to 9}
\item{negmood1}{Negative Mood: Ranging from 1.8 to 5}
\item{children}{Total Number of Children: Ranging from 0-10}
\item{str_chd}{Stress Related to Children: Ranging from 0-8}
...
}
}
\source{
internally created toy dataset
}
\usage{
happyparent
}
\description{
A dataset containing the stress rating of parents and children along with the parents negative mood.
}
\keyword{datasets}
