% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stalker.R
\docType{data}
\name{stalker}
\alias{stalker}
\title{Toy dataset of Therapeutic intervention for stalkers.}
\format{
A data frame with 78 rows and 3 variables:
\describe{
\item{Therapy}{Therapy Type: Cruel to be Kind Therapy, Psychodyshamic Therapy, None}
\item{AmtTher}{Amount of Therapy: 6 Weeks, 12 Weeks}
\item{Stalker}{Time Spent Stalking: Hours per Week}
...
}
}
\source{
internally created toy dataset
}
\usage{
stalker
}
\description{
A dataset containing the therapy type, amount of therapy, and stalking time in hours.
}
\keyword{datasets}
